Refactor dupes
